cloud/aws-dev_course/web_app/requirements.txt
boto3
Flask
Flask-Migrate
Flask-SQLAlchemy
gunicorn
psycopg2-binary
py-healthcheck
python-dotenv
PyYAML
